Update Hime with the code from the VPS
Her life and modifications were always done directly on the VPS because I was too lazy and nano on linux has syntax highlighting which is nice.

This is the last of the code, merged with the current code. The code HAS been edited, tokens removed and overall all the hacks I used to make it use less RAM and be faster were removed to make the code look like code (not decent, this is isn't decent yet).